arrumando erros, e classes

+++ src/main/java/br/com/Sublimoon/PI/entity/Produto.java
@@ -111,7 +111,7 @@ public class Produto  extends AbstractEntity {
 
     @Getter @Setter
     @Column(name = "quantidadeProCarrinho")
-    private float quantidadeProCarrinho;
+    private int quantidadeProCarrinho;
